Upgrade gradlew, Improve InstantStop, improve SprintReset.
