+++ build.gradle
@@ -1,6 +1,6 @@
 plugins {
     id "java"
-    id("com.github.weave-mc.weave-gradle") version "649dba7468"
+    id("com.github.weave-mc.weave-gradle") version "fac948db7f"
 }
 
 group = "legitish"
